Refactor: common responses: align with unleash + added type safety
This change updates the common responses file to bring it in line with
how it's done in unleash/unleash. In addition to removing redundant type
annotations, it also uses the type system to constrain what responses
you can ask for. In other words, you can no longer ask for a response
that doesn't exist.