fix: don't throw errors if someone tries to look up missing toggles
DISCLAIMER: I don't know how best to test this without spinning up a
whole instance and doing proper e2e testing, but I did find a bug
where if you ask the proxy about toggles that don't exist, it'll throw
a 500 because it tries to access the `impressionData` property of a
toggle that doesn't exist (easy to test via curl or the swagger ui).

This change adds a conditional in case `impressionData` is undefined
or the toggle itself is undefined.