fix: OpenAPI POST /features payload is wrong (#80)
* fix: type request body in lookup toggles endpoint

* fix: change toggleNames to toggles

* fix: update openAPI to use toggles instead of toggleNames

* fix: remove non-existing properties

These properties don't exist on the target type, so the compiler was complaining

* refactor:  add test for invalid json queries to /features

* fix: update openAPI description

* fix: update snapshot

* Refactor: common responses: align with unleash + added type safety

This change updates the common responses file to bring it in line with
how it's done in unleash/unleash. In addition to removing redundant type
annotations, it also uses the type system to constrain what responses
you can ask for. In other words, you can no longer ask for a response
that doesn't exist.

* Feat: add bad request standard response.

The POST endpoints may end up not accepting the data you pass them.
This response takes care of that.

* Feat: add bad request responses to POST endpoints

* Fix: add types back, because the _tests_ complain otherwise.

The application runs fine, though

* Fix: handle more errors than just openAPI errors

* fix: update tests to check for malformed json + more non-standard

* fix: don't throw errors if someone tries to look up missing toggles

DISCLAIMER: I don't know how best to test this without spinning up a
whole instance and doing proper e2e testing, but I did find a bug
where if you ask the proxy about toggles that don't exist, it'll throw
a 500 because it tries to access the `impressionData` property of a
toggle that doesn't exist (easy to test via curl or the swagger ui).

This change adds a conditional in case `impressionData` is undefined
or the toggle itself is undefined.

* Fix: allow any properties of the post endpoint to be empty

* Chore: remove duplicated error handling assignment

* fix: specify payload type for test

* fix: add required object properties

I previously removed them as my compiler said they were invalid. Turns
out that it was most likely just due to outdated packages on my end.

* chore: enable swagger UI when running local dev.

This is a suggestion, but I would expect us always to want the UI when
running in dev mode.

* Feat: test that the proxy accepts top-level and nested custom props

* fix: make additional properties optional

* fix: cast context to Context

* Apply suggestions from code review

Co-authored-by: sighphyre <liquidwicked64@gmail.com>

* Update src/openapi/openapi-service.ts

Co-authored-by: sighphyre <liquidwicked64@gmail.com>