chore: add types prop to tsconfig.json
This change adds a "types" property to tsconfig.json pointing at the
openapi types declaration we use in this project.

During development, it seems to be hit or miss whether the language
server actually finds this type declaration, so it's hit or miss
whether anything compiles or not. This addition fixes that issue