docs: integrate docs.io content into readme (#116)
## What

This PR merges the content from docs.getunleash.io with the content that already existed in the readme. In doing so, some of the content has been reordered and potentially reworded.

It also replaces the old connection diagram with a new and snazzier one.

## Why

This is in preparation to do for the proxy what we did for SDKs in [Unleash PR #2858](https://github.com/Unleash/unleash/pull/2858) (and following PRs): reduce two separate information sources into a single document and an updating replica. 

## Commits

* docs: Update readme intro

* docs: add figure

* docs: fix badges with links

* docs: add image

* docs: move configuration higher up

* docs: move api section with content

* docs: add health endpoint; custom strats

* docs: update how to connect and how to run

* docs: remove docusaurus admonition

* docs: use new image for connection overview

* docs: update image caption

* docs: add sdk capabilities section

As mentioned in https://github.com/Unleash/unleash/pull/3060