docs: fix endpoint description POST /proxy
## What

This change updates the operation description for the `POST /proxy`
operation. It changes it from implying that only some toggles are
returned to explicitly saying that all toggles are returned.

## Why

The original description was misleading. Probably because the
author (me 🙋) didn't quite understand how the endpoint worked when
they wrote it.