Docs: update openapi for POST endpoints
## What

This change updates the description of POST `/proxy` and `/proxy/all` and GET `/proxy/all`

It states clearly what happens if you don't provide the optional `toggles` part of the request payload.

It also removes what appears to be copy-paste errors that have been left in.

## Why

Because the information contained within this new bit of documentation was unknown to me and confusing when I came across it.