0.15.2 (#128)
This PR releases v0.15.2. It contains a bugfix (#126) for the recently released content-type checking that the proxy does.

## The content-type bug

Prior to this fix, the content-type checking would apply to all endpoints on the app you use for the proxy. If you use the proxy as a standalone app, then that doesn't really matter, but if you pass in a separate app, then all non-proxy endpoints will also have the content-type checking applied.

The fix, provided by @Dzixxx, makes it so that the content-type validation only happens on the endpoints bundled by the proxy.